fix: rename func and remove package
